[part2] courseinfo exercise 2.5
